move styles into invite modal folder
