+++ site/components/pages/Pledge/InvitationModals/AcceptModal/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 
 import { ButtonPrimary, ButtonSecondary, Text } from "@klimadao/lib/components";
 import { Modal } from "components/Modal";
-import * as styles from "../../PledgeDashboard/styles";
+import * as styles from "../styles";
 import { t, Trans } from "@lingui/macro";
 import {
   approveSecondaryWallet,
