+++ site/components/pages/Pledge/InvitationModals/styles.ts
@@ -1,2 +1,20 @@
+import { css } from "@emotion/css";
 
-
+export const modalButtons = css`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  justify-items: space-between;
+  gap: 1.2rem;
+`;
+export const modalMessage = css`
+  padding: 2rem 0;
+`;
+export const removeTitleContainer = css`
+  display: flex;
+  padding: 2rem 0;
+  gap: 1.2rem;
+`;
+export const lockIcon = css`
+  padding-top: 0.4rem;
+`;
