+++ site/components/pages/Pledge/PledgeDashboard/styles.ts
@@ -22,24 +22,3 @@ export const column = css`
     grid-column: span 1;
   }
 `;
-
-export const modalButtons = css`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  justify-items: space-between;
-  gap: 1.2rem;
-`;
-
-export const modalMessage = css`
-  padding: 2rem 0;
-`;
-export const lockIcon = css`
-  padding-top: 0.4rem;
-`;
-
-export const removeTitleContainer = css`
-  display: flex;
-  padding: 2rem 0;
-  gap: 1.2rem;
-`;
